# docs/schemas/openapi/admin.spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amp Admin API",
...
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/datasets/{namespace}/{name}": {
      "delete": {
        "tags": [
          "datasets"
        ],
        "summary": "Handler for the `DELETE /datasets/{namespace}/{name}` endpoint",
        "description": "Removes all manifest links and version tags for a dataset.\n\n## Path Parameters\n- `namespace`: Dataset namespace\n- `name`: Dataset name\n\n## Response\n- **204 No Content**: Dataset successfully deleted (or didn't exist)\n- **400 Bad Request**: Invalid path parameters\n- **500 Internal Server Error**: Database operation error\n\n## Error Codes\n- `INVALID_PATH`: Invalid namespace or name in path parameters\n- `UNLINK_DATASET_MANIFESTS_ERROR`: Failed to unlink dataset manifests from dataset store\n\n## Behavior\nThis endpoint deletes all metadata for a dataset including:\n- All manifest links in the dataset_manifests table\n- All version tags (cascaded automatically via foreign key constraint)\n- Orphaned manifest files (manifests not referenced by any other dataset)\n\nThis operation is fully idempotent - it returns 204 even if the dataset\ndoesn't exist. Manifests that are still referenced by other datasets are\npreserved.",
        "operationId": "delete_dataset",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Dataset namespace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Dataset name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Dataset successfully deleted"
          },
          "400": {
            "description": "Invalid path parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/datasets/{namespace}/{name}/versions/{revision}/jobs": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Handler for the `GET /datasets/{namespace}/{name}/versions/{revision}/jobs` endpoint",
        "description": "Retrieves and returns all jobs for a specific dataset revision.\n\n## Path Parameters\n- `namespace`: Dataset namespace\n- `name`: Dataset name\n- `revision`: Dataset revision (version, hash, \"latest\", or \"dev\")\n\n## Response\n- **200 OK**: Returns all jobs for the dataset\n- **400 Bad Request**: Invalid path parameters\n- **404 Not Found**: Dataset or revision not found\n- **500 Internal Server Error**: Database connection or query error\n\n## Error Codes\n- `INVALID_PATH`: Invalid path parameters (namespace, name, or revision malformed)\n- `DATASET_NOT_FOUND`: Dataset revision does not exist\n- `RESOLVE_REVISION_ERROR`: Failed to resolve dataset revision (database error)\n- `LIST_JOBS_ERROR`: Failed to list jobs from metadata database (database error)\n- `LIST_JOB_DESCRIPTORS_ERROR`: Failed to list job descriptors from metadata database\n\nThis handler:\n- Validates and extracts the dataset reference from the URL path\n- Resolves the revision to a manifest hash using the dataset store\n- Queries all jobs filtered by the manifest hash\n- Returns all matching jobs",
        "operationId": "list_dataset_jobs",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Dataset namespace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Dataset name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revision",
            "in": "path",
            "description": "Revision (version, hash, latest, or dev)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dataset or revision not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Handler for the `GET /jobs/{id}` endpoint",
        "description": "Retrieves and returns a specific job by its ID from the metadata database.\n\n## Path Parameters\n- `id`: The unique identifier of the job to retrieve (must be a valid JobId)\n\n## Response\n- **200 OK**: Returns the job information as JSON\n- **400 Bad Request**: Invalid job ID format (not parseable as JobId)\n- **404 Not Found**: Job with the given ID does not exist\n- **500 Internal Server Error**: Database connection or query error\n\n## Error Codes\n- `INVALID_JOB_ID`: The provided ID is not a valid job identifier\n- `JOB_NOT_FOUND`: No job exists with the given ID\n- `GET_JOB_ERROR`: Failed to retrieve job from scheduler (database error)\n- `GET_DESCRIPTOR_ERROR`: Failed to retrieve job descriptor from metadata database",
        "operationId": "jobs_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved job information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid job ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Handler for the `DELETE /jobs/{id}` endpoint",
        "description": "Deletes a job by its ID if it's in a terminal state (Completed, Stopped, or Failed).\nThis is a safe, idempotent operation that only removes finalized jobs from the system.\n\n## Path Parameters\n- `id`: The unique identifier of the job to delete (must be a valid JobId)\n\n## Response\n- **204 No Content**: Job was successfully deleted or does not exist (idempotent)\n- **400 Bad Request**: Invalid job ID format (not parseable as JobId)\n- **409 Conflict**: Job exists but is not in a terminal state (cannot be deleted)\n- **500 Internal Server Error**: Database error occurred\n\n## Error Codes\n- `INVALID_JOB_ID`: The provided ID is not a valid job identifier\n- `JOB_CONFLICT`: Job exists but is not in a terminal state\n- `GET_JOB_ERROR`: Failed to retrieve job from scheduler (database error)\n- `DELETE_JOB_ERROR`: Failed to delete job from scheduler (database error)\n\n## Idempotent Behavior\nThis handler is idempotent - deleting a non-existent job returns 204 (success).\nThis allows clients to safely retry deletions without worrying about 404 errors.\n\n## Behavior\nThis handler provides safe job deletion with the following characteristics:\n- Only jobs in terminal states (Completed, Stopped, Failed) can be deleted\n- Non-terminal jobs are protected from accidental deletion\n- Non-existent jobs return success (idempotent behavior)\n- Database layer ensures atomic deletion\n\n## Terminal States\nJobs can only be deleted when in these states:\n- Completed → Safe to delete\n- Stopped → Safe to delete\n- Failed → Safe to delete\n\nProtected states (cannot be deleted):\n- Scheduled → Job is waiting to run\n- Running → Job is actively executing\n- StopRequested → Job is being stopped\n- Stopping → Job is in process of stopping\n- Unknown → Invalid state",
        "operationId": "jobs_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Job deleted successfully or does not exist (idempotent)"
          },
          "400": {
            "description": "Invalid job ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Job cannot be deleted (not in terminal state)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response returned by the API\n\nThis struct represents error information returned in HTTP error responses.\nIt provides structured error details including a machine-readable error code\nand human-readable message.\n\n## Error Code Conventions\n- Error codes use SCREAMING_SNAKE_CASE (e.g., `DATASET_NOT_FOUND`)\n- Codes are stable and can be relied upon programmatically\n- Messages may change and should only be used for display/logging\n\n## Example JSON Response\n```json\n{\n  \"error_code\": \"DATASET_NOT_FOUND\",\n  \"error_message\": \"dataset 'eth_mainnet' version '1.0.0' not found\"\n}\n```",
        "required": [
          "error_code",
          "error_message"
        ],
        "properties": {
          "error_code": {
            "type": "string",
            "description": "Machine-readable error code in SCREAMING_SNAKE_CASE format\n\nError codes are stable across API versions and should be used\nfor programmatic error handling. Examples: `INVALID_SELECTOR`,\n`DATASET_NOT_FOUND`, `METADATA_DB_ERROR`"
          },
          "error_message": {
            "type": "string",
            "description": "Human-readable error message\n\nMessages provide detailed context about the error but may change\nover time. Use `error_code` for programmatic decisions."
          }
        }
      },
...
      "JobInfo": {
        "type": "object",
        "description": "Represents job information for the API response",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "node_id",
          "status",
          "descriptor"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "description": "Job creation timestamp in ISO 8601 / RFC 3339 format"
          },
          "descriptor": {
            "description": "Job descriptor containing job-specific parameters as JSON"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier for the job (64-bit integer)"
          },
          "node_id": {
            "type": "string",
            "description": "ID of the worker node this job is scheduled for"
          },
          "status": {
            "type": "string",
            "description": "Current status of the job (Scheduled, Running, Completed, Stopped, Failed, etc.)"
          },
          "updated_at": {
            "type": "string",
            "description": "Job last update timestamp in ISO 8601 / RFC 3339 format"
          }
        }
      },
...
      "JobsResponse": {
        "type": "object",
        "description": "API response containing job information",
        "required": [
          "jobs"
        ],
        "properties": {
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobInfo"
            },
            "description": "List of jobs"
          },
          "next_cursor": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Cursor for the next page of results (None if no more results)"
          }
        }
      },
...
    }
  },
...
}